models/code/requirements.txt
boto3
pillow
numpy
